+++ CHANGELOG.md
@@ -8,7 +8,12 @@
 
 - Prefix mod ID to initialize messages.
 - Add logging for dev features.
+- Switch to QM on intermediary.
+
+### Removals
+
+- Removed JamLib config reload command.
 
 Closed issues: None.
 
-[Full Changelog](https://github.com/JamCoreModding/JamLib/compare/0.4.4...0.4.5)
+[Full Changelog](https://github.com/JamCoreModding/JamLib/compare/0.4.4...0.5.0)